Support .slnf file formats
