+++ src/Cli/dotnet/commands/dotnet-test/SolutionAndProjectUtility.cs
@@ -1,11 +1,11 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
+using Microsoft.Build.Evaluation;
 using Microsoft.DotNet.Tools;
-using Microsoft.DotNet.Tools.Test;
-using Microsoft.VisualStudio.SolutionPersistence;
-using Microsoft.VisualStudio.SolutionPersistence.Model;
-using Microsoft.VisualStudio.SolutionPersistence.Serializer;
+using Microsoft.DotNet.Tools.Common;
+
+using LocalizableStrings = Microsoft.DotNet.Tools.Test.LocalizableStrings;
 
 namespace Microsoft.DotNet.Cli
 {
@@ -21,21 +21,21 @@ namespace Microsoft.DotNet.Cli
                 return (false, string.Format(LocalizableStrings.CmdNonExistentDirectoryErrorDescription, directory));
             }
 
-            var possibleSolutionPaths = GetSolutionFilePaths(directory);
+            var solutionPaths = GetSolutionFilePaths(directory);
 
             // If more than a single sln file is found, an error is thrown since we can't determine which one to choose.
-            if (possibleSolutionPaths.Length > 1)
+            if (solutionPaths.Length > 1)
             {
                 return (false, string.Format(CommonLocalizableStrings.MoreThanOneSolutionInDirectory, directory));
             }
 
-            if (possibleSolutionPaths.Length == 1)
+            if (solutionPaths.Length == 1)
             {
-                var possibleProjectPaths = GetProjectFilePaths(directory);
+                var projectPaths = GetProjectFilePaths(directory);
 
-                if (possibleProjectPaths.Length == 0)
+                if (projectPaths.Length == 0)
                 {
-                    projectOrSolutionFilePath = possibleSolutionPaths[0];
+                    projectOrSolutionFilePath = solutionPaths[0];
                     isSolution = true;
                     return (true, string.Empty);
                 }
@@ -44,16 +44,32 @@ namespace Microsoft.DotNet.Cli
             }
             else  // If no solutions are found, look for a project file
             {
-                string[] possibleProjectPath = GetProjectFilePaths(directory);
+                string[] projectPaths = GetProjectFilePaths(directory);
 
-                if (possibleProjectPath.Length == 0)
+                if (projectPaths.Length == 0)
                 {
-                    return (false, LocalizableStrings.CmdNoProjectOrSolutionFileErrorDescription);
+                    var solutionFilterPaths = GetSolutionFilterFilePaths(directory);
+
+                    if (solutionFilterPaths.Length == 0)
+                    {
+                        return (false, LocalizableStrings.CmdNoProjectOrSolutionFileErrorDescription);
+                    }
+
+                    if (solutionFilterPaths.Length == 1)
+                    {
+                        projectOrSolutionFilePath = solutionFilterPaths[0];
+                        isSolution = true;
+                        return (true, string.Empty);
+                    }
+                    else
+                    {
+                        return (false, LocalizableStrings.CmdMultipleProjectOrSolutionFilesErrorDescription);
+                    }
                 }
 
-                if (possibleProjectPath.Length == 1)
+                if (projectPaths.Length == 1)
                 {
-                    projectOrSolutionFilePath = possibleProjectPath[0];
+                    projectOrSolutionFilePath = projectPaths[0];
                     return (true, string.Empty);
                 }
 
@@ -68,39 +84,68 @@ namespace Microsoft.DotNet.Cli
                 .ToArray();
         }
 
+        private static string[] GetSolutionFilterFilePaths(string directory)
+        {
+            return Directory.EnumerateFiles(directory, CliConstants.SolutionFilterExtensionPattern, SearchOption.TopDirectoryOnly).ToArray();
+        }
+
         private static string[] GetProjectFilePaths(string directory) => [.. Directory.EnumerateFiles(directory, CliConstants.ProjectExtensionPattern, SearchOption.TopDirectoryOnly).Where(IsProjectFile)];
 
         private static bool IsProjectFile(string filePath) => CliConstants.ProjectExtensions.Contains(Path.GetExtension(filePath), StringComparer.OrdinalIgnoreCase);
 
-        public static async Task<IEnumerable<string>> ParseSolution(string solutionFilePath, string directory)
+        public static string GetRootDirectory(string solutionOrProjectFilePath)
         {
-            if (string.IsNullOrEmpty(solutionFilePath))
+            string fileDirectory = Path.GetDirectoryName(solutionOrProjectFilePath);
+            return string.IsNullOrEmpty(fileDirectory) ? Directory.GetCurrentDirectory() : fileDirectory;
+        }
+
+        public static IEnumerable<Module> GetProjectPropertiesInternal(string projectFilePath, ProjectCollection projectCollection)
+        {
+            var project = projectCollection.LoadProject(projectFilePath);
+            return ExtractModulesFromProject(project);
+        }
+
+        private static IEnumerable<Module> ExtractModulesFromProject(Project project)
+        {
+            _ = bool.TryParse(project.GetPropertyValue(ProjectProperties.IsTestProject), out bool isTestProject);
+
+            if (!isTestProject)
             {
-                VSTestTrace.SafeWriteTrace(() => $"Solution file path cannot be null or empty: {solutionFilePath}");
                 return [];
             }
 
-            var projectsPaths = new List<string>();
-            SolutionModel solution = null;
+            _ = bool.TryParse(project.GetPropertyValue(ProjectProperties.IsTestingPlatformApplication), out bool isTestingPlatformApplication);
 
-            try
+            string targetFramework = project.GetPropertyValue(ProjectProperties.TargetFramework);
+            string targetFrameworks = project.GetPropertyValue(ProjectProperties.TargetFrameworks);
+            string targetPath = project.GetPropertyValue(ProjectProperties.TargetPath);
+            string projectFullPath = project.GetPropertyValue(ProjectProperties.ProjectFullPath);
+            string runSettingsFilePath = project.GetPropertyValue(ProjectProperties.RunSettingsFilePath);
+
+            var projects = new List<Module>();
+
+            if (string.IsNullOrEmpty(targetFrameworks))
             {
-                solution = SolutionSerializers.GetSerializerByMoniker(solutionFilePath) is ISolutionSerializer serializer
-                    ? await serializer.OpenAsync(solutionFilePath, CancellationToken.None)
-                    : null;
+                projects.Add(new Module(targetPath, PathUtility.FixFilePath(projectFullPath), targetFramework, runSettingsFilePath, isTestingPlatformApplication, isTestProject));
             }
-            catch (Exception ex)
+            else
             {
-                VSTestTrace.SafeWriteTrace(() => $"Failed to parse solution file '{solutionFilePath}': {ex.Message}");
-                return [];
+                var frameworks = targetFrameworks.Split(CliConstants.SemiColon, StringSplitOptions.RemoveEmptyEntries);
+                foreach (var framework in frameworks)
+                {
+                    project.SetProperty(ProjectProperties.TargetFramework, framework);
+                    project.ReevaluateIfNecessary();
+
+                    projects.Add(new Module(project.GetPropertyValue(ProjectProperties.TargetPath),
+                        PathUtility.FixFilePath(projectFullPath),
+                        framework,
+                        runSettingsFilePath,
+                        isTestingPlatformApplication,
+                        isTestProject));
+                }
             }
 
-            if (solution is not null)
-            {
-                projectsPaths.AddRange(solution.SolutionProjects.Select(project => Path.Combine(directory, project.FilePath)));
-            }
-
-            return projectsPaths;
+            return projects;
         }
     }
 }
