+++ test/dotnet-test.Tests/GivenDotnetTestBuildsAndRunsTestsWithArtifacts.cs
@@ -54,11 +54,6 @@ namespace Microsoft.DotNet.Cli.Test.Tests
 
             // Read MSTestVersion from Versions.props and update the .csproj file
             // Search for Versions.props file from the current directory up to the root
-
-            Console.WriteLine($"CurrentDirectory: {Directory.GetCurrentDirectory()}");
-            Console.WriteLine($"BaseDirectory : {AppContext.BaseDirectory}");
-            Console.WriteLine($"testInstance.Path : {testInstance.Path}");
-            Console.WriteLine($"TestContext.Current.TestExecutionDirectory : {TestContext.Current.TestExecutionDirectory}");
             string? versionsPropsPath = PathUtility.FindFileInParentDirectories(TestContext.Current.TestExecutionDirectory, $"eng{Path.DirectorySeparatorChar}Versions.props") ?? throw new FileNotFoundException("Versions.props file not found.");
             string msTestVersion = testInstance.ReadMSTestVersionFromProps(versionsPropsPath);
             testInstance.UpdateProjectFileWithMSTestVersion(Path.Combine($@"{testInstance.Path}{PathUtility.GetDirectorySeparatorChar()}TestProject", "TestProject.csproj"), msTestVersion);
